chore: Rename Category
PEIT-30 Category Carousel
